refacto headers organization
