+++ source/parser/filename_expansion.c
@@ -10,7 +10,6 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/42sh.h"
 
 // GROS CHANTIER ATTENTION
