+++ source/utils/utils_files.c
@@ -10,8 +10,12 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../include/42sh.h"
 #include "utils.h"
+#include "libft.h"
+#include "exec.h"
+
+#include <stdlib.h>
+#include <unistd.h>
 
 char *get_next_line(int fd) {
 	char buffer[2];
